+++ language/duchain/codemodel.cpp
@@ -93,7 +93,7 @@ class CodeModelRepositoryItem {
     return file.index();
   }
   
-  size_t itemSize() const {
+  uint itemSize() const {
     return dynamicSize();
   }
   
@@ -122,7 +122,7 @@ class CodeModelRequestItem {
     return m_item.hash();
   }
   
-  size_t itemSize() const {
+  uint itemSize() const {
       return m_item.itemSize();
   }
